server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django>=5.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
twilio==9.9.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
python-dotenv
